requirements.txt
psycopg2==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
